requirements.txt
pandas==0.25.1
numpy
google-cloud-core==1.0.3
google-cloud-storage==1.19.0
sendgrid==6.1.0
bocfx